pyproject.toml
[tool.poetry]
name = "hexlet_code"
version = "0.1.0"
description = "Show difference betwene two files"
authors = ["Anton <zavr35rus@gmail.com>"]
packages = [
    { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest = "^6.1.2"
coverage = "^5.3"

[tool.poetry.scripts]
gendiff="gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
